% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{as.character}
\alias{as.character}
\alias{as.character.LAW}
\title{Prints an object of class LAW}
\usage{
\method{as.character}{LAW}(x, ...)
}
\arguments{
\item{x}{An object of \code{class} \code{LAW}.}

\item{\dots}{Not used.}
}
\value{
A character string  recalling what can be  done with an  object of
  \code{class} \code{LAW}.
}
\description{
Prints an object of \code{class} \code{LAW}.
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
Other methods and functions for LAW objects: 
\code{\link{LAW}()},
\code{\link{alter}()},
\code{\link{copy}()},
\code{\link{evaluate_eic}()},
\code{\link{evaluate_psi}()},
\code{\link{evaluate_remainder}()},
\code{\link{get_feature}()},
\code{\link{reveal}()},
\code{\link{sample_from}()}
}
\concept{methods and functions for LAW objects}
