% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compute_Gbar_hatW}
\alias{compute_Gbar_hatW}
\title{Computes the conditional probability of A = 1 given W=W_i}
\usage{
compute_Gbar_hatW(W, Gbar_hat, threshold = 0.05)
}
\arguments{
\item{W}{A \code{vector} of real numbers between 0 and 1.}

\item{Gbar_hat}{The output  of \code{estimate_Gbar}, derived by training an
algorithm on a learning data set (see '?estimate_Gbar').}

\item{threshold}{A \code{numeric} taking values between 0 and 1/2 (defaults
to 0.05).}
}
\value{
A \code{vector}  containing  'GW'  defined as  \eqn{Gbar_hat(W)}
  bounded away from 0 and 1  by 'threshold', where \eqn{Gbar_hat(W)} is the
  conditional probability that A=1 given W predicted by the fit 'Gbar_hat'.
}
\description{
Given  realizations of  W  in [0,1]  and an  estimator  of the  conditional
probability  that   A=1  given  W  (as   output  by  \code{estimate_Gbar}),
\code{compute_Gbar_hatW}  computes  the  conditional probability  that  A=1
given W=W_i.
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
Other estimating functions: \code{\link{compute_Qbar_hatAW}},
  \code{\link{compute_lGbar_hatAW}}, \code{\link{wrapper}}
}
\concept{estimating functions}
