% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compute_gcomp}
\alias{compute_gcomp}
\title{Builds the G-computation estimator}
\usage{
compute_gcomp(QW, Qbar, nobs)
}
\arguments{
\item{QW}{The marginal  law of  'W', a  \code{tibble} with  columns named
'value' and 'weight' (a discrete law) as output by \code{estimate_QW}.}

\item{Qbar}{The  conditional  expectation of 'Y' given '(A,W)', a \code{function}.}

\item{nobs}{The sample size of the data set used to estimate the above 'QW'
and  'Qbar', an  \code{integer}.  Used  only  for the  estimation of  the
standard deviation of the G-computation  estimator, which is not reliable.}
}
\value{
A  \code{tibble} containing  the  value  of the  G-computation  estimator
  ('psi_n'  column) and  that of  the estimator  of its  standard deviation
  ('sig_n' column).
}
\description{
Given a marginal law  of W and a conditional expectation  of Y given (A,W),
the so called 'GW' and 'Qbar' features  of the law of an experiment, either
a  priori known  or estimated  beforehand, \code{compute_gcomp}  builds the
G-computation estimator  of \eqn{Psi}  at that law.
}
\details{
Caution:  the estimator  of the  standard deviation  of the  G-computation
  estimator can be trusted only in very specific circumstances.
}
\examples{

## create an experiment and draw a data set from it
example(tlrider, echo = FALSE)
obs <- sample_from(experiment, n = 250)

## estimate 'QW' and 'Qbar'
QW_hat <- estimate_QW(obs)
Qbar_hat <- estimate_Qbar(obs, working_model_Q_one)

## wrap 'Qbar_hat' (a fit) into a function
Qbar_hat_fun <- wrapper(Qbar_hat, FALSE)

## compute the G-computation estimator
(compute_gcomp(QW_hat, Qbar_hat_fun, nrow(obs)))

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
\code{\link{estimate_QW}}  to  estimate the  marginal  law of  W,
  \code{\link{estimate_Qbar}} to estimate  the conditional expectation of Y
  given  (A,W), \code{\link{compute_iptw}}  to compute the  IPTW estimator,
  \code{\link{wrapper}}.
}
