% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compute_iptw}
\alias{compute_iptw}
\title{Builds the IPTW estimator}
\usage{
compute_iptw(dat, Gbar, threshold = 0.05)
}
\arguments{
\item{dat}{The learning data  set. Must have the  same form as a  data set
produced    by    an    object   of    \code{class}    \code{LAW}    (see
'?tlrider').}

\item{Gbar}{The  conditional  probability  that  'A=1'   given  'W',  a
\code{function}.}

\item{threshold}{A small  positive number (default  value 0.05) used  to bound
'Gbar' away from zero.}
}
\value{
A  \code{tibble} containing  the  value  of the  IPTW  estimator
  ('psi_n'  column) and  that of  the estimator  of its  standard deviation
  ('sig_n' column).
}
\description{
Given a data set  consisting of realizations of (W,A,Y) in  [0,1] x {0,1} x
[0,1] drawn  from a  common law  and the  conditional probability  that A=1
given W, the so called 'Gbar' feature  of the law, either a priori known or
estimated  beforehand, \code{compute_iptw}  builds  the  IPTW estimator  of
\eqn{Psi} at the law that generated the data set.
}
\details{
Caution: when the 'Gbar'  feature is estimated, the  estimator of
  the  standard deviation  of  the IPTW  estimator  is conservative,  i.e.,
  over-estimates its target.
}
\examples{

## create an experiment and draw a data set from it
example(tlrider, echo = FALSE)
obs <- sample_from(experiment, n = 250)

## suppose that 'Gbar' is known
Gbar <- get_feature(experiment, "Gbar")

## compute the IPTW estimator
(compute_iptw(obs, Gbar))

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
\code{\link{estimate_Gbar}}   to  estimate   the   conditional
  probability that A=1 given  W, \code{\link{compute_gcomp}} to compute the
  G-computation estimator.
}
