% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compute_lGbar_hatAW}
\alias{compute_lGbar_hatAW}
\title{Computes the conditional likelihood of A_i given W_i}
\usage{
compute_lGbar_hatAW(A, W, Gbar_hat, threshold = 0.05)
}
\arguments{
\item{A}{A \code{vector} of 0 and 1.}

\item{W}{A \code{vector} of real numbers between 0 and 1.}

\item{Gbar_hat}{The output  of \code{estimate_Gbar}, derived by training an
algorithm on a learning data set (see '?estimate_Gbar').}

\item{threshold}{A \code{numeric} taking values between 0 and 1/2 (defaults
to 0.05).}
}
\value{
A \code{vector} containing  'lGAW' defined as the  maximum between
  'threshold' and\deqn{A  * Gbar_hat(W)  + (1  - A)  * (1  - Gbar_hat(W)),}
  where \eqn{Gbar_hat(W)} is  the conditional probability that  A=1 given W
  predicted by the fit 'Gbar_hat'.
}
\description{
Given  realizations of  (W,A) in  [0,1]  x {0,1}  and an  estimator of  the
conditional    probability   that    A=1    given   W    (as   output    by
\code{estimate_Gbar}), \code{compute_lGbar_hatAW}  computes the conditional
probability that A=A_i given W=W_i.
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2018).
}
\seealso{
Other estimating functions: \code{\link{compute_Qbar_hatAW}},
  \code{\link{wrapper}}
}
\concept{estimating functions}
