% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compute_Qbar_hatAW}
\alias{compute_Qbar_hatAW}
\title{Computes the conditional expectation of Y given (A,W)=(A_i,W_i)}
\usage{
compute_Qbar_hatAW(A, W, Qbar_hat, blip = FALSE)
}
\arguments{
\item{A}{A \code{vector} of 0 and 1.}

\item{W}{A \code{vector} of real numbers between 0 and 1.}

\item{Qbar_hat}{The output  of \code{estimate_Qbar}, derived by training an
algorithm on a learning data set (see '?estimate_Qbar').}

\item{blip}{A \code{logical}  (defaults to 'FALSE') indicating  whether or
not to output the so called blip function (see below).}
}
\value{
A \code{vector}  containing either \eqn{Qbar_hat(A,W)}  (ig 'blip'
  argument  set to  'FALSE') or  \eqn{Qbar_hat(1,W) -  Qbar_hat(0,W)} where
  \eqn{Qbar_hat(A,W)}  is  the conditional  expectation  of  Y given  (A,W)
  predicted by the fit 'Qbar_hat'.
}
\description{
Given  realizations of  (W,A) in  [0,1]  x {0,1}  and an  estimator of  the
conditional    expectation   of    Y    given   (A,W)    (as   output    by
\code{estimate_Qbar}),  \code{compute_Qbar_hatAW} computes  the conditional
expectation of Y  given (A,W)=(A_i,W_i).
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
Other estimating functions: 
\code{\link{compute_Gbar_hatW}()},
\code{\link{compute_lGbar_hatAW}()},
\code{\link{wrapper}()}
}
\concept{estimating functions}
