% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{copy}
\alias{copy}
\alias{copy.LAW}
\title{Copies an object of class LAW}
\usage{
\method{copy}{LAW}(this, ...)
}
\arguments{
\item{this}{An object of \code{class} \code{LAW}.}

\item{...}{Not used.}
}
\value{
A copy of argument \code{this}.
}
\description{
Given an  object of class  \code{LAW}, makes a copy  of it in  another such
object.
}
\examples{

## create an experiment, make a copy (once)
example(tlrider, echo = FALSE)
experiment_bis <- experiment
experiment_ter <- copy(experiment)
identical(experiment, experiment_bis)
identical(experiment, experiment_ter)

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2018).
}
\seealso{
Other methods and functions for LAW objects: \code{\link{LAW}},
  \code{\link{alter}}, \code{\link{as.character}},
  \code{\link{evaluate_eic}}, \code{\link{evaluate_psi}},
  \code{\link{evaluate_remainder}},
  \code{\link{get_feature}}, \code{\link{reveal}},
  \code{\link{sample_from}}
}
\concept{methods and functions for LAW objects}
