% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_Gbar}
\alias{estimate_Gbar}
\title{Estimates the Gbar feature of an object of class LAW}
\usage{
estimate_Gbar(dat, algorithm, ...)
}
\arguments{
\item{dat}{The learning data  set. Must have the  same form as a  data set
produced    by    an    object   of    \code{class}    \code{LAW}    (see
'?tlrider').}

\item{algorithm}{The  algorithm.  See the  built-in algorithms  by running
'data(algorithms)'.}

\item{...}{Additional parameters passed to the algorithm.}
}
\value{
The  result of  the fit  output by  the algorithm  trained on  the
  learning data set.
}
\description{
Given a data set  consisting of realizations of (W,A,Y) in  [0,1] x {0,1} x
[0,1]  drawn   from  a  common  law,   \code{estimate_Gbar}  estimates  the
conditional probability that  A=1 given W, the so called  'Gbar' feature of
the law.
}
\examples{

## create an experiment and draw a data set from it
example(tlrider, echo = FALSE)
obs <- sample_from(experiment, n = 250)

## load the built-in algorithms
data(algorithms)
working_model_G_one

## estimate 'Gbar'
Gbar_hat <- estimate_Gbar(obs, working_model_G_one)

}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
\seealso{
\code{\link{estimate_QW}} to estimate the marginal distribution of
  W, \code{\link{estimate_Qbar}} to estimate the conditional expectation of
  Y given (A,W).
}
