% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_QW}
\alias{estimate_QW}
\title{Estimates the marginal law of W non-parametrically}
\usage{
estimate_QW(dat)
}
\arguments{
\item{dat}{The learning data  set. Must have the  same form as a  data set
produced    by    an    object   of    \code{class}    \code{LAW}    (see
'?tlrider').}
}
\value{
A  \code{tibble}  with  columns named  'value'  and 'weight'  that
  describes the empirical law of 'W' in data set 'dat'.
}
\description{
Given a data set  consisting of realizations of (W,A,Y) in  [0,1] x {0,1} x
[0,1]  drawn   from  a   common  law,  \code{estimate_QW}   estimates  non-
parametrically the marginal law of W, the so called 'QW' feature.
}
\examples{

## create an experiment and draw a data set from it
example(tlrider, echo = FALSE)
obs <- sample_from(experiment, n = 250)

## estimate 'QW'
QW_hat <- estimate_QW(obs)


}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2018).
}
\seealso{
\code{\link{estimate_Gbar}}   to  estimate   the   conditional
  probability that A=1 given W, \code{\link{estimate_Qbar}} to estimate the
  conditional expectation of Y given (A,W).
}
