% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{evaluate_remainder}
\alias{evaluate_remainder}
\title{Evaluates the remainder term at a law w.r.t. another law}
\usage{
evaluate_remainder(this, that, params = NULL)
}
\arguments{
\item{this}{An object of \code{class} \code{LAW} describing, say, \eqn{P}.}

\item{that}{An object of \code{class} \code{LAW} describing, say, \eqn{P'}.}

\item{params}{If not  \code{NULL} (default value), a \code{list} consisting
of  two  \code{list}s,  each specifying  additional  parameters  possibly
needed to fully characterize the corresponding law.}
}
\value{
A  \code{numeric},  the value of the remainder term \eqn{Rem_P(P')}.
}
\description{
Consider two laws \eqn{P} and \eqn{P'} for (W,A,Y) in [0,1] x {0,1} x [0,1]
characterized by  two objects of  \code{class} \code{LAW}. Recall  that the
mapping \eqn{\Psi} and  its efficient influence curve at  \eqn{P} are given
by \deqn{\Psi(P) =  E_P {Qbar(1,W) - Qbar(0,W)}} with  \eqn{Qbar(A,W) = E_P
(Y|A,W)}   and   \eqn{D^*    (P)   =   D_1^*   (P)    +   D_2^*(P)}   where
\deqn{D_1^*(P)(W,A,Y)   =   Qbar(1,W)   -    Qbar(0,W)   -   \Psi(P)}   and
\deqn{D_2^*(P)(W,A,Y) = \frac{2A-1}{Gbar(A,W)} (Y - Qbar(A,W)).}
}
\details{
This  function evaluates  the remainder  term at  \eqn{P'} w.r.t.  \eqn{P}:
\deqn{Rem_P(P') = \Psi(P') - \Psi(P) - (P - P')D^*(P').}

The evaluation can be performed if one  has built the two objects or if one
acts as an  \emph{oracle}.  Works only if all the  relevant features of the
laws are characterized.
}
\examples{

## create a one-dimensional fluctutation (collection of laws)
example(tlrider, echo = FALSE)
(evaluate_psi(another_experiment, h = 0)) # Psi(P_0)
(evaluate_psi(another_experiment, h = 1)) # Psi(P_1)

## evaluate the remainder term
## (hint: same 'Gbar' feature, so ...)
evaluate_remainder(another_experiment, another_experiment,
                   list(list(h = 0), list(h = 1)))

## create another one-dimensional fluctuation
yet_another_experiment <- copy(another_experiment)
alter(yet_another_experiment, Gbar = function(w){rep_len(1/2, length(w))})

## evaluate the remainder term
evaluate_remainder(another_experiment, yet_another_experiment,
                   list(list(h = 0), list(h = 1)))

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
\code{\link{evaluate_psi}} to  evaluate \eqn{Psi}  at a  law and
  \code{\link{evaluate_eic}} to  evaluate the efficient influence  curve of
  \eqn{Psi} at a law.

Other methods and functions for LAW objects: \code{\link{LAW}},
  \code{\link{alter}}, \code{\link{as.character}},
  \code{\link{copy}}, \code{\link{evaluate_eic}},
  \code{\link{evaluate_psi}}, \code{\link{get_feature}},
  \code{\link{reveal}}, \code{\link{sample_from}}
}
\concept{methods and functions for LAW objects}
