% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fluctuate}
\alias{fluctuate}
\title{Fluctuates a Qbar feature of an object of class LAW}
\usage{
fluctuate(Qbar, Gbar, h)
}
\arguments{
\item{Qbar}{A   conditional  expectation  of   'Y'  given   '(A,W)',  a
\code{function}.  Must  have  the  same  form  as  the  Qbar  feature  of
'experiment' (see '?get_feature').}

\item{Gbar}{A  conditional   probability  that  'A=1'  given   'W',  a
\code{function}. Must  have  the  same  form  as  the  Gbar  feature  of
'experiment' (see '?get_feature').}

\item{h}{A \code{numeric}}
}
\value{
A \code{function}.
}
\description{
Fluctuates a  Qbar feature  \eqn{\bar{Q}} of  an object of  class LAW  in a
direction characterized  by a  Gbar feature \eqn{\bar{G}}  of an  object of
class  LAW.   Specifically,  for  any  real  number  \eqn{h},  \deqn{\logit
(\bar{Q}_h(A,W)) =  \logit (\bar{Q}_h(A,W)) + h  \frac{2A-1}{A \bar{G}(W) +
(1-A) (1 - \bar{G}(W))}.}
}
\examples{

## create an experiment, sample from it
example(tlrider, echo = FALSE)
experiment
obs <- sample_from(experiment, n = 250)

## extract its Gbar and Qbar features
Gbar <- get_feature(experiment, "Gbar")
Qbar <- get_feature(experiment, "Qbar")

## fluctuate Qbar using Gbar
Q_h <- fluctuate(Qbar, Gbar, h = 1/2)
Q_h(cbind(A=0:1, W = c(1/3, 2/3)))

## estimate 'Gbar' and 'Qbar'
Gbar_hat <- estimate_Gbar(obs, working_model_G_one)
Qbar_hat <- estimate_Qbar(obs, working_model_Q_one)

## wrap 'Gbar_hat' and 'Qbar_hat' (two fits) into two functions
Gbar_hat_fun <- wrapper(Gbar_hat, FALSE)
Qbar_hat_fun <- wrapper(Qbar_hat, FALSE)

## fluctuate Qbar_hat_fun using Gbar_hat_fun
Qhat_h <- fluctuate(Qbar_hat_fun, Gbar_hat_fun, h = 1/2)
Qhat_h(cbind(A=0:1, W = c(1/3, 2/3)))

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
\code{\link{estimate_Qbar}}   to  estimate   the   conditional
  expectation of Y given (A,W), \code{\link{estimate_Gbar}} to estimate the
  conditional probability that A=1 given (A,W), \code{\link{wrapper}}.
}
