% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{get_feature}
\alias{get_feature}
\alias{get_feature.LAW}
\title{Reveals one feature of an object of class LAW}
\usage{
\method{get_feature}{LAW}(this, what, ...)
}
\arguments{
\item{this}{An object of \code{class} \code{LAW}.}

\item{what}{A \code{character}, one among "Qbar", "Gbar", "QW", "qY".}

\item{\dots}{Additional parameters  possibly needed to  fully characterize
the law.}
}
\value{
Depending on  argument  'what',  one among  \itemize{\item  'QW',
  marginal law of  'W', a \code{function} (the density)  or a \code{tibble}
  with columns  named 'value'  and 'weight'  (a discrete  law)\item 'Gbar',
  conditional  probability that  'A=1' given  'W', a  \code{function} \item
  'Qbar', conditional  mean of 'Y'  given '(A,W)', a  \code{function} \item
  'qY', conditional density of 'Y' given '(A,W)', a \code{function}.}  Each
  of them equals 'NA' if it is not characterized in object 'this'.
}
\description{
Reveals  one  feature   of  an  object  of   \code{class}  \code{LAW}  that
characterizes a law for (W,A,Y) in [0,1]  x {0,1} x [0,1]. Can be performed
if one has built the object or if one acts as an \emph{oracle}.
}
\examples{

## create then reveal an experiment
example(tlrider, echo = FALSE)
get_feature(experiment, "Gbar")


}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2018).
}
\seealso{
\code{\link{reveal}} to reveal  some relevant features of the law.

Other methods and functions for LAW objects: \code{\link{LAW}},
  \code{\link{alter}}, \code{\link{as.character}},
  \code{\link{copy}}, \code{\link{evaluate_eic}},
  \code{\link{evaluate_psi}},
  \code{\link{evaluate_remainder}}, \code{\link{reveal}},
  \code{\link{sample_from}}
}
\concept{methods and functions for LAW objects}
