% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{label}
\alias{label}
\title{Labels the elements of a data set}
\format{
An object of class \code{function} of length 1.
}
\usage{
label
}
\arguments{
\item{xx}{A \code{vector}.}

\item{sample_size}{A \code{vector} of \code{integer}s.}
}
\value{
A \code{vector} of \code{characters}.
}
\description{
Labels the elements of a data set to prepare subsamples.
}
\details{
Let 'L'  be the length of 'sample_size' and  'Q'be the quotient in
  the Euclidean division of the  length of 'xx' by 'sum(sample_size)'.  The
  \code{function} \code{label}  creates 'Q  x L'  different strings  of the
  form 'q_l'.   Each 'q_l' string  is repeated 'sample_size[l]'  times. The
  function is used to make easily sub-data sets of a data set.
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\keyword{datasets}
