% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{reveal}
\alias{reveal}
\alias{reveal.LAW}
\title{Reveals an object of class LAW}
\usage{
\method{reveal}{LAW}(this, ...)
}
\arguments{
\item{this}{An object of \code{class} \code{LAW}.}

\item{\dots}{Additional parameters  possibly needed to  fully characterize
the law.}
}
\value{
A \code{list} with tags\itemize{\item  'QW', marginal law of 'W', a
  \code{function}  (the  density) or  a  \code{tibble}  with columns  named
  'value'   and  'weight'   (a  discrete   law)\item  'Gbar',   conditional
  probability  that  'A=1'  given  'W',  a  \code{function}  \item  'Qbar',
  conditional  mean of  'Y' given  '(A,W)', a  \code{function} \item  'qY',
  conditional  density  of  'Y'  given  '(A,W)',  a  \code{function}  \item
  'sample_from', a \code{function}  to sample from the law.}   Each of them
  equals 'NA' if it is not characterized.
}
\description{
Reveals an object  of \code{class} \code{LAW} by exhibiting  a selection of
relevant features of the law for (W,A,Y)  in [0,1] x {0,1} x [0,1] that the
object characterizes.  Can be performed if  one has built the  object or if
one acts as an \emph{oracle}.
}
\examples{

## create then reveal an experiment
example(tlrider, echo = FALSE)
reveal(experiment)


}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
\seealso{
\code{\link{sample_from}}  to sample from the law (if  it is fully
  characterized), \code{\link{alter}}  to modify some relevant  features of
  the law, \code{\link{evaluate_psi}} to evaluate the value of \eqn{Psi} at
  the  law,  and  \code{\link{evaluate_eic}}   to  evaluate  the  efficient
  influence curve of \eqn{Psi} at the law.

Other methods and functions for LAW objects: \code{\link{LAW}},
  \code{\link{alter}}, \code{\link{as.character}},
  \code{\link{copy}}, \code{\link{evaluate_eic}},
  \code{\link{evaluate_psi}},
  \code{\link{evaluate_remainder}},
  \code{\link{get_feature}}, \code{\link{sample_from}}
}
\concept{methods and functions for LAW objects}
