% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{sample_from}
\alias{sample_from}
\alias{sample_from.LAW}
\title{Samples from an object of class LAW}
\usage{
\method{sample_from}{LAW}(this, n = 1, ...)
}
\arguments{
\item{this}{An object of \code{class} \code{LAW}.}

\item{n}{An \code{integer}, the wished sample size. Defaults to 'n = 1'.}

\item{\dots}{Additional parameters  possibly needed to  fully characterize
the law.}
}
\value{
Either  \code{NULL} if  the law  is undetermined  or the  data set
  sampled, a \code{tibble} or \code{data.frame} with columns named 'W', 'A'
  and 'Y'.
}
\description{
Samples data of the form (W,A,Y) in [0,1] x {0,1} x [0,1] from an object of
\code{class} \code{LAW} when it fully characterizes a law.
}
\examples{

## create an experiment
example(tlrider, echo = FALSE)
experiment

## sample three observations from it
sample_from(experiment, n = 3)

}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\seealso{
\code{\link{reveal}} to reveal  some relevant features of the law,
  \code{\link{alter}}   to  modify   them,  \code{\link{evaluate_psi}}   to
  evaluate    the     value    of    \eqn{Psi}    at     the    law,    and
  \code{\link{evaluate_eic}} to  evaluate the efficient influence  curve of
  \eqn{Psi} at the law.

Other methods and functions for LAW objects: 
\code{\link{LAW}()},
\code{\link{alter}()},
\code{\link{as.character}()},
\code{\link{copy}()},
\code{\link{evaluate_eic}()},
\code{\link{evaluate_psi}()},
\code{\link{evaluate_remainder}()},
\code{\link{get_feature}()},
\code{\link{reveal}()}
}
\concept{methods and functions for LAW objects}
