% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{sample_from_mixture_of_uniforms}
\alias{sample_from_mixture_of_uniforms}
\title{Samples from a mixture of uniform laws}
\format{
An object of class \code{function} of length 1.
}
\usage{
sample_from_mixture_of_uniforms
}
\arguments{
\item{n}{An \code{integer}, the sample size.}

\item{mixture_weights}{A  \code{vector} of \code{numerics}  that specifies
the weights given to the uniform laws. Must sum up to one.}

\item{mins}{A  \code{vector}   of  \code{numerics}  that  specifies  the
lower-bounds of the supports of the uniform laws.}

\item{maxs}{A  \code{vector}   of  \code{numerics}  that  specifies  the
upper-bounds of the supports of the uniform laws.}
}
\value{
A \code{vector} of numerics, the sample.
}
\description{
Samples  from the  mixture of  uniform laws  characterized by  the provided
weights and corresponding intervals.
}
\references{
Benkeser & Chambaz, "A Ride in Targeted Learning Territory" (2019).
}
\keyword{datasets}
