% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{wrapper}
\alias{wrapper}
\title{A convenient wrapper to make predictions based on a fit}
\usage{
wrapper(fit)
}
\arguments{
\item{fit}{A fit as output by \code{estimate_Gbar} or \code{estimate_Qbar}.}
}
\value{
A function to make predictions.
}
\description{
A convenient wrapper to make predictions  based on a fit produced by either
\code{estimate_Gbar} or \code{estimate_Qbar}.
}
\seealso{
Other estimating functions: \code{\link{compute_Gbar_hatW}},
  \code{\link{compute_Qbar_hatAW}},
  \code{\link{compute_lGbar_hatAW}}
}
\concept{estimating functions}
